requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
python-dotenv
langchain_chroma
sentence-transformers
langchain-huggingface